+++ src/main.py
@@ -1,6 +1,7 @@
 from menus.main_menu import run_main_menu
 from menus.file_menu import run_file_menu
 from menus.text_menu import run_text_menu
+from menus.csv_menu  import run_csv_menu
 
 def main():
     while True:
@@ -10,6 +11,8 @@ def main():
             run_text_menu()
         elif pilihan == "2":
             run_file_menu()
+        elif pilihan == "3":
+            run_csv_menu()
         elif pilihan == "0":
             print("Keluar dari program...")
             break