+++ src/menus/main_menu.py
@@ -12,6 +12,7 @@ def run_main_menu():
 =================\n
 1.Text Tools
 2.File Tools
+3.CSV  Tools
 0. EXIT
     """)
     choice = input("Pilih :")